[CH-TESTS] whether homepage is fetched correctly
