+++ app/templates/index.html
@@ -53,6 +53,7 @@
     <!-- scripts -->
     <script type="text/javascript" src="{{ url_for('static', filename='js/jquery-3.2.1.min.js') }}"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='js/app.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='js/welcome.js') }}"></script>
     <script src="https://unpkg.com/expect@%3C21/umd/expect.min.js"></script>
 
 </body>
